backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
google-genai==1.10.0
chromadb==0.5.0
pydantic==2.9.0
python-docx==1.1.0
PyPDF2==3.0.1
numpy==1.26.4
python-dotenv==1.0.1
